Update chat_to_files.py
I ran into issues where i was moving the file around and renaming it but renamed it back before i launched gpt engineer and ran into it an error because the file was not in the file_dict. This will address that. I am also making a comment in this issue https://github.com/gpt-engineer-org/gpt-engineer/issues/928